commit_3

+++ Reena/src/main/java/com/restaurantmanagement/Reena/service/CustomerService.java
@@ -19,15 +19,34 @@ public class CustomerService {
         return customerRepository.findAll();
     }
 
-    public Optional<Customer> findById(Long id) {
-        return customerRepository.findById(id);
+    public Customer findById(Long id) {
+        Optional<Customer> customerOptional = customerRepository.findById(id);
+        return customerOptional.orElse(null);
     }
 
     public Customer save(Customer customer) {
         return customerRepository.save(customer);
     }
 
-    public void deleteById(Long id) {
-        customerRepository.deleteById(id);
+    public Customer update(Long id, Customer customerDetails) {
+        Optional<Customer> optionalCustomer = customerRepository.findById(id);
+        if (optionalCustomer.isPresent()) {
+            Customer existingCustomer = optionalCustomer.get();
+            existingCustomer.setFirstName(customerDetails.getFirstName());
+            existingCustomer.setLastName(customerDetails.getLastName());
+            existingCustomer.setEmail(customerDetails.getEmail());
+            // Set other fields as needed
+            return customerRepository.save(existingCustomer);
+        }
+        return null;
+    }
+
+    public boolean deleteById(Long id) {
+        Optional<Customer> optionalCustomer = customerRepository.findById(id);
+        if (optionalCustomer.isPresent()) {
+            customerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
